pyproject.toml
[project]
name = "chainlit-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.17",
    "python-dotenv>=1.1.0",
]
